bots/search_predict_bot/requirements.txt
google-cloud-storage==2.6.0
emoji==1.7.0
nltk==3.7
regex==2022.10.31
keras==2.9.0
Keras-Preprocessing==1.1.2
tensorflow==2.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
oauthlib==3.2.1
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.10.1
urllib3==1.26.12